Remove sample.lua from data files
The file is kept around in the pandoc-lua-engine test-suite.

Closes: #8356
